Parse MP3 audio data to fragment shader as texture
Currently, this is just test phase parsing mp3 audio data and parsing it
into a fragment shader (which can be visualised with
shaders/audio.glsl). If there is no audio supplied, a sine wave will be
in substitute, will change it so that's not the case later. This will all definitely need to be cleaned up and
changed a bit